Software engineering project
This project was done in college using Visual Studio 2022 and C#. The task was to create a program that would store data for energy monitoring of public buildings. In this project it was enough to do data manipulation, reports and login together with data protection for each user. 

The reason this program would not work is because it was taken from my collage account and I don't have access to the database used in the project on this account.